:truck: mv: move codes to the FanningImages component
